% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{awesome_list}
\alias{awesome_list}
\title{Fetch and parse an Awesome List README.md file}
\usage{
awesome_list(url, github.basic.info = FALSE)
}
\arguments{
\item{url}{URL for Github repository}

\item{github.basic.info}{Logical; fetch additional basic info about the repo}
}
\value{
Summarized data frame
}
\description{
Dowloads an Awesome List and transform it to the form of easy-to-use data.frame.
Optionally, the function may fetch some additional data describing libraries hosted on Github.
}
