% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_curves.R
\name{clusterGearyCurves}
\alias{clusterGearyCurves}
\title{Cluster Geary's C curves}
\usage{
clusterGearyCurves(
  sfes,
  hclust_params = list(),
  leiden_params = list(resolution = 0.8, objective_function = "modularity")
)
}
\arguments{
\item{sfes}{A list of \code{SpatialFeatureExperiment} objects which have the
same genes, same sample IDs, and with Moran's I computed for the genes. The
names of the list must be the bin size.}

\item{hclust_params}{Hierarchical clustering parameters, passed to
\code{\link[bluster]{HclustParam}}.}

\item{leiden_params}{Leiden clustering parameters, passed to
\code{\link[igraph]{cluster_leiden}}.}
}
\value{
A data frame.
}
\description{
Just like \code{\link{clusterMoranCurves}}, but for Geary's C.
}
