% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_curves.R
\name{clusterLeeCurves}
\alias{clusterLeeCurves}
\title{Cluster Lee's L curves}
\usage{
clusterLeeCurves(
  df,
  hclust_params = list(),
  leiden_params = list(resolution = 0.8, objective_function = "modularity"),
  mat = NULL
)
}
\arguments{
\item{df}{Data frame with Lee's L results from \code{\link{runBinAnalyses}},
which should have columns pair, side, and lee. The data frame can be read
into R with \code{\link{readLeeSamples}}, where a cutoff can be set to
remove gene pairs with low Lee's L in all bin sizes.}

\item{hclust_params}{Hierarchical clustering parameters, passed to
\code{\link[bluster]{HclustParam}}.}

\item{leiden_params}{Leiden clustering parameters, passed to
\code{\link[igraph]{cluster_leiden}}.}

\item{mat}{Matrix with column as bin side lengths and rows as genes. If NULL,
then it will be made from \code{df}.}

\item{sides}{Numeric vector of bin sizes whose results are to be read.}
}
\value{
The same data frame in the input but with cluster assignment of each
  gene pair added.
}
\description{
Lee's L has been computed for pairs of genes across spatial scales. This
function reads the results and clusters the curves of Lee's L of each pair of
genes across scales. The \code{\link[bluster]{approxSilhouette}} function can
be used to assess cluster quality.
}
