% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_curves.R
\name{clusterMoranCurves}
\alias{clusterMoranCurves}
\title{Cluster Moran's I curves}
\usage{
clusterMoranCurves(
  df,
  hclust_params = list(),
  leiden_params = list(resolution = 0.8, objective_function = "modularity"),
  mat = NULL
)
}
\arguments{
\item{df}{Data frame from \code{\link{runBinAnalyses}} for Moran's I, with
columns moran, gene, and side.}

\item{hclust_params}{Hierarchical clustering parameters, passed to
\code{\link[bluster]{HclustParam}}.}

\item{leiden_params}{Leiden clustering parameters, passed to
\code{\link[igraph]{cluster_leiden}}.}

\item{mat}{Matrix with column as bin side lengths and rows as genes. If NULL,
then it will be made from \code{df}.}
}
\value{
The same data frame in the input but with cluster assignment of each
gene added.
}
\description{
Moran's I has been computed for the same genes across different spatial
scales. This function gets the Moran's I values for all genes and resolutions
and clusters the patterns with which Moran's I's vary through scales, as a
way to cluster genes. Leiden and hierarchical clustering are performed on the
Moran's I values themselves and on the diff between adjacent scales. The
\code{\link[bluster]{approxSilhouette}} function can be used to assess
cluster quality.
}
