% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aggregates.R
\name{getBinOverlapProp}
\alias{getBinOverlapProp}
\title{Find portion of each bin overlapping the tissue or cells}
\usage{
getBinOverlapProp(
  sfe,
  tissue_geometry,
  BPPARAM = SerialParam(),
  batch_size = 1000,
  prop = TRUE
)
}
\arguments{
\item{sfe}{SFE object with the aggregated bins, generated with
\code{\link{makeAggregates}}.}

\item{tissue_geometry}{Either tissue boundary (with holes if present in
tissue) or cell segmentation polygons. Area of overlap of each bin with
this geometry will be computed.}

\item{BPPARAM}{A \code{\link[BiocParallel]{bpparam}}, to parallelize over
batches of bins.}

\item{batch_size}{Number of bins in each batch.}

\item{prop}{Logical, whether to return proportions of bin area in tissue
instead of actual area.}
}
\value{
A numeric vector same length as \code{ncol(sfe)}.
}
\description{
If tissue boundary is used, then it should allow for holes, because holes can
also cause edge effect.
}
