% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getCellChatDB}
\alias{getCellChatDB}
\title{Download CellChat database}
\usage{
getCellChatDB(species = c("human", "mouse"), bfc = BiocFileCache())
}
\arguments{
\item{species}{Either human or mouse}

\item{bfc}{\code{\link[BiocFileCache]{BiocFileCache}} instance where you can
set where the files will be stored.}
}
\value{
A list holding the database
}
\description{
Taken from [CellChat's source
code](https://github.com/jinworks/CellChat/tree/main/data), to use without
installing CellChat, which can't be a imported or suggested by this package
because CellChat is not on CRAN or Bioconductor.
}
