% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat.R
\name{getCellChatInfo}
\alias{getCellChatInfo}
\title{Filter Lee's L results for gene pairs in CellChatDB}
\usage{
getCellChatInfo(
  df_lees_lmm,
  genes,
  species = c("human", "mouse"),
  bfc = BiocFileCache()
)
}
\arguments{
\item{df_lees_lmm}{Output for Lee's L from \code{\link{runBinLMM}}}

\item{genes}{A vector of gene symbols for the genes to consider; for
smFISH-based technology, the list of gene of interest is often much smaller
than the number of genes in the database.}

\item{species}{Either "human" or "mouse".}

\item{bfc}{A \code{\link[BiocFileCache]{BiocFileCache}} instance for where
the CellChat database will be downloaded.}
}
\value{
A data frame with the LMM results and additional columns from
  CellChat database for more info about the interaction.
}
\description{
This function takes CellChat's database without using the CellChat package
itself. It filter the Lee's L LMM results (see \code{\link{runBinLMM}}) to
only include gene pairs in the database for known interactions, and adds info
about the interactions from the database such as the KEGG pathway ID. Note
that gene pairs that are not in the database can still be interesting because
they tell something about cell neighborhoods.
}
