% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getMoranMeanVar}
\alias{getMoranMeanVar}
\title{Get the mean and variance of Moran's I}
\usage{
getMoranMeanVar(sfes, name = c("moran", "geary"))
}
\arguments{
\item{sfes}{A list of SFE objects of different bin sizes from the same sample
(can be read with \code{\link{readBins}}), required to compute and plot the
mean and variance of Moran's I (or Geary's C). The list must have names
corresponding to bin sizes.}

\item{name}{Whether mean and variance are to be computed for Moran's I (moran)
or Geary's C (geary).}
}
\value{
A data frame with columns side, mean, var, th1, and th2. Column th1
  is the 2.5% percentile in the normal distribution given mean and var, and
  th2 is the 97.5% percentile, after Bonferroni correction for the number of
  genes and number of bin sizes so the interval is rather conservative.
}
\description{
This function computes the mean and variance of Moran's I under the null
hypothesis of no spatial autocorrelation, in the context of plotting Moran's
I curves over bin sizes in \code{\link{plotMoranCurves}}.
}
\seealso{
[spdep::moran.test()]
}
