% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lmm.R
\name{getPredCurves}
\alias{getPredCurves}
\title{Get fitted curves per group}
\usage{
getPredCurves(df_res)
}
\arguments{
\item{df_res}{Data frame output from \code{\link{runBinLMM}}, with list columns "model" and "data".}
}
\value{
A data frame with column "pred" for the predicted curves.
}
\description{
After running \code{\link{runBinLMM}} with \code{save_models = TRUE}, use this function
to get the predicted curves for each group.
}
