% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-api.R
\name{github_extract_user_repo}
\alias{github_extract_user_repo}
\title{Extract user name and repo name from a Github repo URL}
\usage{
github_extract_user_repo(repo.url)
}
\arguments{
\item{repo.url}{A Github repo.url, character}
}
\value{
List with two values: user.name and repo.name
}
\description{
Extract user name and repo name from a Github repo URL
}
\examples{
github_extract_user_repo("https://github.com/strengejacke/sjstats")
}
