% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-api.R
\name{github_readme_url}
\alias{github_readme_url}
\title{Get link to raw README.md file}
\usage{
github_readme_url(url)
}
\arguments{
\item{url}{Github repo URL}
}
\value{
Dowload URL to README.md
}
\description{
This functions checks, if URL is not a link to the raw Github README.md file.
If doesn't, it extracts user and repo names from the URL and then fetches README.md URL using Github API v3.
}
\examples{
github_readme_url("https://github.com/qinwf/awesome-r/")
}
