% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-api.R
\name{github_repo_basic_info}
\alias{github_repo_basic_info}
\title{Get basic info about Github repo}
\usage{
github_repo_basic_info(repo.url, user = NA, repo = NA)
}
\arguments{
\item{user}{character username}

\item{repo}{character repo name}
}
\value{
Combined basic info from two functions: `github_repo_info` and `github_topics`
}
\description{
Get basic info about Githb repo, i. e. number of stars, number of forks and topics.
}
\examples{
github_repo_basic_info("github.com/tidyverts/tsibble")
}
