% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-api.R
\name{github_repo_info}
\alias{github_repo_info}
\title{Get basic information about Github repo}
\usage{
github_repo_info(repo.url, user = NA, repo = NA)
}
\arguments{
\item{user}{character username}

\item{repo}{character repo name}
}
\value{
List of Github repo attributes
}
\description{
Get all repo info offered by the main type of Gihub API requests.
}
\examples{
github_repo_info("github.com/qinwf/awesome-r")
}
\references{
For detailed info see: https://developer.github.com/v3/
}
