% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{github_topics}
\alias{github_topics}
\title{Get Github Topics for given username and repo}
\usage{
github_topics(repo.url, user = NA, repo = NA)
}
\arguments{
\item{user}{character username}

\item{repo}{character repo name}
}
\value{
List of Github Topics which describe the repo
}
\description{
Get Github Topics using Github API v3.
}
\examples{
github_topics("easystats", "see")
}
