% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-api.R
\name{is_github}
\alias{is_github}
\title{Check, if URL is a Github link}
\usage{
is_github(url)
}
\arguments{
\item{url}{Github repo URL}
}
\value{
Logical value: if URL is from Github?
}
\description{
This functions returns true if link leads us to github.com or raw.githubusercontent.com
}
\examples{
is_github("https://github.com/qinwf/awesome-r/")
}
