% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_control.R
\name{makeMonteCarlo}
\alias{makeMonteCarlo}
\title{Create Monte Carlo simulations with no spatial structure}
\usage{
makeMonteCarlo(
  in_path,
  out_path,
  tissue_boundary,
  quantiles = c(0.25, 0.5, 0.75),
  nsim = 100,
  BPPARAM = SerialParam(),
  queen = FALSE
)
}
\arguments{
\item{in_path}{Path with SFE objects written to subdirectories with
\code{\link{runBinAnalyses}}.}

\item{out_path}{Output directory}

\item{tissue_boundary}{\code{sf} or \code{sfc} for tissue boundary. See
\code{\link[SpatialFeatureExperiment]{getTissueBoundaryConcave}} and
\code{\link[SpatialFeatureExperiment]{getTissueBoundaryImg}} on obtaining
the tissue boundary.}

\item{quantiles}{Vector of quantiles of total counts of genes to use for the
intensity of the Poisson point process, because edge effect affects genes
with different expression levels differently.}
}
\value{
\code{out_path}; the output is written to disk there.
}
\description{
The purpose of this function is to check if edge effect is causing spurious
changes in Moran's I and Lee's L across scales.This function uses geometries
from existing aggregations to create Monte Carlo simulations of gene
expression with no spatial structure. Poisson samples with mean informed by
the total counts of a given gene are taken on small bins, which are then
aggregated. Because given complete spatial randomness (CSR), the transcript
spots should be a homogeneous Poisson point process within the tissue
boundary. The number of counts within any given area follows a Poisson
distribution whose mean is the mean of the Poisson point process multiplied
by the size of that area. Therefore, to make computation faster, we can take
Poisson samples with means scaled by the area of bins in tissue.
}
\details{
This function creates uncorrelated non-spatial patterns. They are stored as
SFE objects and written to disk. Then spatial metrics can be computed on
those patterns. This can be used to check if edge effect is causing artifacts
in the multi-scale analysis, and to compute Monte Carlo p-values for the
linear mixed models used to compare across biological conditions where the
log likelihood ratio test p-values are unreliable due to correlation of
spatial metrics between adjacent bin sizes and heteroscedasticity, i.e.
variance is higher for larger bins because there are fewer bins.

Because edge effect artifacts affect genes with different expression levels
differently, \code{nsim} CSR patterns are generated for each quantile of
expression levels.
}
