% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Piezo1-download}
\alias{Piezo1-download}
\alias{Piezo1TxSpots}
\alias{Piezo1TissueBoundary}
\alias{Piezo1Binned}
\alias{Piezo1BinAnalyses}
\alias{Piezo1Xenium}
\title{Download Xenium data from Xue et al.}
\usage{
Piezo1TxSpots(
  sample = c("all", "expanded1", "expanded2", "expanded1_pi7", "expanded2_pi7",
    "nonexpanded1", "nonexpanded2", "yoda1", "yoda2"),
  bfc = BiocFileCache()
)

Piezo1TissueBoundary(
  sample = c("all", "expanded1", "expanded2", "expanded1_pi7", "expanded2_pi7",
    "nonexpanded1", "nonexpanded2", "yoda1", "yoda2"),
  bfc = BiocFileCache()
)

Piezo1Binned(
  sample = c("all", "expanded1", "expanded2", "expanded1_pi7", "expanded2_pi7",
    "nonexpanded1", "nonexpanded2", "yoda1", "yoda2"),
  bfc = BiocFileCache()
)

Piezo1BinAnalyses(
  sample = c("all", "expanded1", "expanded2", "expanded1_pi7", "expanded2_pi7",
    "nonexpanded1", "nonexpanded2", "yoda1", "yoda2"),
  bfc = BiocFileCache()
)

Piezo1Xenium(
  sample = c("all", "expanded1", "expanded2", "expanded1_pi7", "expanded2_pi7",
    "nonexpanded1", "nonexpanded2", "yoda1", "yoda2"),
  bfc = BiocFileCache()
)
}
\arguments{
\item{sample}{Which sample(s) to download, can have length greater than 1 to
download multiple samples. If "all", then all samples will be downloaded.
See \code{\link{sample_info}} for information about each sample.}

\item{bfc}{\code{\link[BiocFileCache]{BiocFileCache}} instance where you can
set where the files will be stored.}
}
\value{
A character vector of file paths, except for
  \code{Piezo1TissueBoundary} which returns a `sf` data frame.
}
\description{
This dataset is from [The mechanotransducer Piezo1 coordinates metabolism and
inflammation to promote skin
growth](https://www.nature.com/articles/s41467-025-62270-3) by Xue et al. To
make the transcript spot files small enough to download as an example
dataset, I only kept transcripts assigned to cells and only kept the columns
x_location, y_location, cell_id, and feature_name.
}
\details{
These are the functions to download different aspects of the  data:
\describe{
\item{Piezo1TxSpots}{Transcript spot coordinates in csv.gz files}
\item{Piezo1TissueBoundary}{Tissue boundary of each sample in RDS files;
here the files will be read into R as a `sf` data frame.}
\item{Piezo1Binned}{Original binned data without processing or analyses in tar.gz}
\item{Piezo1BinAnalyses}{Binned data after removing edge bins, and with PCA,
Moran's I, and Lee's L results}
}
}
