% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{plotEigenvalues}
\alias{plotEigenvalues}
\title{Plot variance explained across bin sizes}
\usage{
plotEigenvalues(
  sfes,
  npcs = 20,
  reduction = "PCA",
  field = "varExplained",
  xlab = "Principal component",
  ylab = "Variance explained"
)
}
\arguments{
\item{sfes}{A list of SFE (SCE and SPE are fine) objects with the dimension.
Names of the list are bin sizes.
reduction of interest computed and stored in \code{reducedDims}.}

\item{npcs}{Number of components whose variance explained or eigenvalues are
to be plotted.}

\item{reduction}{Name of the dimension reduction of interest}

\item{field}{Field in the attribute of the dimension reduction where variance
explained or eigenvalues are stored.}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}
}
\value{
A \code{ggplot} object
}
\description{
This function plots the iconic PCA elbow plot for variance explained or
eigenvalue for multiple SCE objects, distinguished by color.
}
