% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLeeCurves}
\alias{plotLeeCurves}
\title{Plot Lee's L curves}
\usage{
plotLeeCurves(df, facet_by = NULL, show_median = FALSE, sample_n = NULL)
}
\arguments{
\item{df}{Data frame output from \code{\link{clusterLeeCurves}}, or any data
frame with a column named "pair" for gene pairs, "side" for bin size as in
side length, and "lee" for Lee's L values, and optionally other categorical
columns indicating cluster membership of genes.}

\item{facet_by}{Name of a categorical or integer column in \code{df} to facet
the plot, not tidyeval.}

\item{show_median}{Logical, whether to plot the median of clusters, only used
when \code{facet_by} is specified.}

\item{sample_n}{A smaller number of gene pairs to plot. Often the number of
gene pairs is large so downsampling makes the plot more readable.}
}
\value{
A \code{ggplot2} object.
}
\description{
The curves show how Lee's L changes over spatial scale for each pair of
genes, as indicated by bin size used to aggregate the data.
}
