% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLeeSelect}
\alias{plotLeeSelect}
\title{Plot select Lee's L curves}
\usage{
plotLeeSelect(df, lmm_res, pairs_use, title = NULL)
}
\arguments{
\item{df}{Data frame with Lee's L results from all samples, with columns
pair, side, lee, sample, and group (for biological conditions).}

\item{lmm_res}{Linear mixed model results data frame from
\code{\link{runBinLMM}}}

\item{pairs_use}{Which pairs to plot}

\item{title}{Title of the plot}
}
\value{
A \code{ggplot2} oboject
}
\description{
Plot Lee's L curves from select gene pairs, coloring by biological conditions
}
