% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{plotLoadingCurves}
\alias{plotLoadingCurves}
\title{Plot gene loadings of a given principal component across scales}
\usage{
plotLoadingCurves(
  sfes,
  pc = 1,
  reduction = "PCA",
  field = "rotation",
  color_by = c("moran", "loading"),
  color_ind = 1,
  flip = FALSE
)
}
\arguments{
\item{sfes}{A list of SFE objects with names as bin sizes}

\item{pc}{Index of principal component or eigenvector or loading vector to
plot}

\item{reduction}{Name of dimension reduction to plot}

\item{field}{Name of the attribute of the dimension reduction that holds the
gene loadings}

\item{color_by}{Whether to color the curves by Moran's I or loading values}

\item{color_ind}{Since one curve can only have one color, choose the index of
the SFE object in \code{sfes} whose values are used for coloring.}

\item{flip}{Logical, whether to flip the signs of the loadings.}
}
\value{
A \code{ggplot} object
}
\description{
This can visualize how the meanings of principal components change with
spatial scales.
}
