% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMoranCurves}
\alias{plotMoranCurves}
\title{Plot Moran's I curves}
\usage{
plotMoranCurves(
  df,
  name = "moran",
  color_by = NULL,
  facet_by = NULL,
  show_null = FALSE,
  mean_vars = NULL,
  color_name = NULL,
  title_name = "Moran's I"
)
}
\arguments{
\item{df}{Data frame output from \code{\link{clusterMoranCurves}}, or any
data frame with a column named "gene" for genes, "side" for bin size as in
side length, and "moran" for Moran's I values, and optionally other
categorical columns indicating cluster membership of genes.}

\item{name}{Column name in \code{df} with the spatial results.}

\item{color_by}{A data frame with column "gene" for gene symbols or IDs and
one other column for color. Alternatively, name of a column in \code{df} to
use for coloring.}

\item{facet_by}{Name of a categorical or integer column in \code{df} to facet
the plot, not tidyeval.}

\item{show_null}{Logical, whether to show expected values of Moran's I under
null hypothesis (values are randomly permuted in space) and the interval
within which the null is not to be rejected (p >= 0.05 after Bonferroni
correction accounting for the number of genes and number of sides).}

\item{mean_vars}{Data frame for means and variances of Moran's I under null
hypothesis from \code{\link{getMoranMeanVar}}, required for \code{show_null
= TRUE}.}

\item{color_name}{Name to show for colors in the legend.}

\item{title_name}{Name of the metric to put on the plot title.}
}
\value{
A \code{ggplot} object.
}
\description{
The curves show how Moran's I changes over spatial scale, as indicated by bin
size used to aggregate the data. Despite the name, this function can be
applied to other univariate spatial results such as Geary's C.
}
