% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSFEsBiscale}
\alias{plotSFEsBiscale}
\title{Plot pairs of genes in space from multiple SFE objects}
\usage{
plotSFEsBiscale(
  sfes,
  pair_use,
  df_lee,
  colGeometryName = 1L,
  bbox = NULL,
  palette = "BlueGold",
  exprs_value = "logcounts",
  swap_rownames = NULL,
  title = NULL,
  crop = TRUE,
  swap_colors = FALSE,
  show_sizes = TRUE,
  sep = "(?<!HLA)-",
  ncol = NULL,
  nrow = NULL,
  samples_use = NULL,
  side_use = NULL,
  widths = NULL,
  heights = NULL,
  design = NULL,
  legend_index = NULL
)
}
\arguments{
\item{sfes}{A list of SFE objects, whose names must be the bin sizes.}

\item{pair_use}{Gene pair to plot}

\item{df_lee}{Data frame from \code{\link{clusterLeeCurves}}, where Lee's L
values are extracted to annotate the plot.}

\item{colGeometryName}{Name of a \code{colGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute the metric. Use
\code{\link{colGeometryNames}} to look up names of the \code{sf} data
frames associated with cells/spots.}

\item{bbox}{Bounding box to plot a subarea}

\item{palette}{Name of bivariate palette; see \code{\link[biscale]{bi_pal}}
for complete list of built-in palettes in the `biscale` package.}

\item{swap_rownames}{Column name of \code{rowData(object)} to be used to
identify features instead of \code{rownames(object)} when labeling plot
elements. If not found in \code{rowData}, then rownames of the gene count
matrix will be used.}

\item{swap_colors}{Logical, whether to swap the positions of the two genes
in `pair_use` in the palette.}

\item{ncol}{Number of columns if plotting multiple features. Defaults to
\code{NULL}, which means using the same logic as \code{facet_wrap}, which
is used by \code{patchwork}'s \code{\link{wrap_plots}} by default.}

\item{widths, heights}{The relative widths and heights of each column and row
in the grid. Will get repeated to match the dimensions of the grid. The
special value of \code{NA}/\verb{-1null} will behave as \verb{1null} unless a fixed aspect
plot is inserted in which case it will allow the dimension to expand or
contract to match the aspect ratio of the content}

\item{design}{Specification of the location of areas in the layout. Can either
be specified as a text string or by concatenating calls to \code{\link[patchwork:area]{area()}} together.
See the examples for further information on use.}
}
\value{
A \code{patchwork} object.
}
\description{
This function is useful to visualize pairs of genes whose Lee's L is of
interest.
}
