% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpca.R
\name{plotWpca}
\alias{plotWpca}
\title{Plot projections of genes/gene pairs in Wpca space}
\usage{
plotWpca(wpca_out, dims = 1:2, color_by = NULL, size = 0.5, label = TRUE)
}
\arguments{
\item{wpca_out}{Output of \code{\link{runWpca}}.}

\item{dims}{Which PCs to plot, as ingeter indices}

\item{color_by}{Vector to color the points, must be in the same order as the
row names of \code{wpca_out$x}}

\item{size}{Point size}
}
\value{
A \code{ggplot2} object. \code{\link[ggforce]{fact_matrix}} will be
  used when plotting more than 2 PCs.
}
\description{
It's similar to \code{scater::plotPCA}, but each point is a gene or gene pair
instead of a cell.
}
