% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpca.R
\name{plotWpcaLoadings}
\alias{plotWpcaLoadings}
\title{Plot Wpca loadings}
\usage{
plotWpcaLoadings(
  wpca_out,
  dims = 1:4,
  nfeatures = 10,
  balanced = TRUE,
  ncol = NULL
)
}
\arguments{
\item{wpca_out}{Output from \code{\link{runWpca}}.}

\item{dims}{Which PCs to plot, as ingeter indices}

\item{nfeatures}{Total number of features whose loadings are to be plotted}

\item{balanced}{Logical, whether to plot equal number of features with the
most positive and most negative loadings.}

\item{ncol}{Passed to \code{\link[ggplot2]{facet_wrap}}, for the number of
columns in the facets when plotting multiple PCs.}
}
\value{
A \code{ggplot2} object
}
\description{
The loadings are for each combination of bin size and group. This helps
interpreting variations in multi-scale behaviors in spatial metrics.
}
