% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readBins}
\alias{readBins}
\title{Read aggregated data from multiple bin sizes for the same sample}
\usage{
readBins(dir, sides)
}
\arguments{
\item{dir}{Directory where the results are stored. The SFE object for each
bin size must be in a directory whose name begins with "binx_esda" where x
is the bin size, such as "bin12_esda".}

\item{sides}{Numeric vector of bin sizes to read}
}
\value{
A list of SFE objects whose names are the bin sizes
}
\description{
This function reads from the output of \code{\link{runBinAnalyses}}. SFE
objects of the selected bin sizes will be loaded.
}
