% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readBinSamples}
\alias{readBinSamples}
\title{Read one bin size from multiple samples}
\usage{
readBinSamples(dirs, side)
}
\arguments{
\item{dirs}{A character vector of directories, one for each sample. The SFE
object for each bin size must be in a directory whose name begins with
"binx_esda" where x is the bin size, such as "bin12_esda".}

\item{side}{One bin size to read}
}
\value{
A list of SFE objects whose names are base names in \code{dirs}.
}
\description{
This function reads from the output of \code{\link{runBinAnalyses}}. Results
from one bin size are read from multiple samples.
}
