% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readLeeSamples}
\alias{readLeeSamples}
\title{Read multi-sample Lee's L results from multiple samples}
\usage{
readLeeSamples(dirs, sample_info = NULL, cutoff = 0L)
}
\arguments{
\item{dirs}{A character vector of directories, one for each sample. There
must be a \code{df_lee.csv} file in each directory.}

\item{sample_info}{Optional data frame with more info about each sample.
There must be a column called "sample". Such info is helpful when plotting.}

\item{cutoff}{Gene pairs whose absolute values of Lee's L is below this
cutoff in all bin sizes and all samples will not be included in the output.
This will remove genes that are not spatially correlated at any length
scale and any sample.}
}
\value{
A data frame with columns moran, side, sample, and optionally other
  columns in \code{sample_info}.
}
\description{
\code{\link{runBinAnalyses}} should write Lee's L for all genes in all bin
sizes to a file \code{df_lee.csv}. This function reads this file from
multiple samples and concatenates them.
}
