% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readMoranSamples}
\alias{readMoranSamples}
\title{Read multi-scale Moran's I results from multiple samples}
\usage{
readMoranSamples(dirs, sample_info = NULL)
}
\arguments{
\item{dirs}{A character vector of directories, one for each sample. There
must be a \code{df_moran.csv} file in each directory.}

\item{sample_info}{Optional data frame with more info about each sample.
There must be a column called "sample". Such info is helpful when plotting.}
}
\value{
A data frame with columns moran, side, sample, and optionally other
  columns in \code{sample_info}.
}
\description{
\code{\link{runBinAnalyses}} should write Moran's I for all genes in all bin
sizes to a file \code{df_moran.csv}. This function reads this file from
multiple samples and concatenates them.
}
