% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpca.R
\name{runWpca}
\alias{runWpca}
\title{Run weighted PCA on fitted curves per stage}
\usage{
runWpca(df, weights_col, ...)
}
\arguments{
\item{df}{A data frame, output from \code{\link{getPredCurves}} or
\code{\link{getEMM}}.}

\item{...}{More arguments to pass to \code{\link[WCluster]{Wpca}}, except for
\code{x} and \code{wcol}.}

\item{df_mean_var}{A dara frame with a column \code{var} for the variance of
the spatial metric for each bin size and each sample, and a column
\code{group} for the groups being compared in the linear mixed models. For
example, output from \code{\link{getMoranMeanVar}}, but with groups added.
See vignette. This data frame is used to compute the weights so the larger
bin sizes and samples with fewer cells get less weight. The weight is
1/var. The maximum weight among all samples in a group is used.}
}
\value{
Same list output as \code{\link[WCluster]{Wpca}}, except that row
  names have been added to the item "rotation" to facilitate further
  visualization.
}
\description{
To further interpret variabilities in the multi-scale behaviors. This is to
be done on fitted curves from \code{\link{getPredCurves}} to better elucidate
trends when there is more variability among individual samples within the
same group. This function is a thin wrapper around
\code{\link[WCluster]{Wpca}}, doing some data wrangling to convert the input
data frame into a matrix as input to that function.
}
