% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_info}
\alias{sample_info}
\title{Sample info for the Piezo1 Xenium data}
\format{
A data frame
}
\source{
https://www.nature.com/articles/s41467-025-62270-3
}
\usage{
sample_info
}
\description{
This is a data frame with two columns: sample and condition. Condition is the
experimental condition. PI14 stands for 14 days post complete injection, and
PI7 is 7 days. Injection means saline was injected into an implant beneath
the skin to expand the skin.
}
\keyword{datasets}
