% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awesome-list.R
\name{awesome_list}
\alias{awesome_list}
\title{Fetch and parse an Awesome List README.md file}
\usage{
awesome_list(url)
}
\arguments{
\item{url}{URL for Github repository}
}
\value{
Summarized data frame
}
\description{
Dowloads an Awesome List and transform it to the form of easy-to-use data.frame.
Optionally, the function may fetch some additional data describing libraries hosted on Github.
}
\examples{
awesome_list('https://github.com/lauris/awesome-scala')
}
