% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_curves.R
\name{clusterLeeCurves}
\alias{clusterLeeCurves}
\title{Cluster Lee's L curves}
\usage{
clusterLeeCurves(
  dir,
  sides,
  cutoff = 0.2,
  hclust_params = list(),
  leiden_params = list(resolution = 0.8, objective_function = "modularity")
)
}
\arguments{
\item{dir}{Directory where Lee's L results are stored. File names must have
the pattern "bin<x>_lee.rds", such as "bin12_lee.rds".}

\item{sides}{Numeric vector of bin sizes whose results are to be read.}

\item{cutoff}{Only gene pairs with Lee's L greater than the cutoff at at
least one bin size are used for clustering, due to the potentially large
number of pairs.}

\item{hclust_params}{Hierarchical clustering parameters, passed to
\code{\link[bluster]{HclustParam}}.}

\item{leiden_params}{Leiden clustering parameters, passed to
\code{\link[igraph]{cluster_leiden}}.}
}
\value{
A data frame.
}
\description{
Lee's L has been computed for pairs of genes across spatial scales. This
function reads the results and clusters the curves of Lee's L of each pair of
genes across scales. The \code{\link[bluster]{approxSilhouette}} function can
be used to assess cluster quality.
}
