% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_curves.R
\name{clusterMoranCurves}
\alias{clusterMoranCurves}
\title{Cluster Moran's I curves}
\usage{
clusterMoranCurves(
  sfes,
  hclust_params = list(),
  leiden_params = list(resolution = 0.8, objective_function = "modularity")
)
}
\arguments{
\item{sfes}{A list of \code{SpatialFeatureExperiment} objects which have the
same genes, same sample IDs, and with Moran's I computed for the genes. The
names of the list must be the bin size.}

\item{hclust_params}{Hierarchical clustering parameters, passed to
\code{\link[bluster]{HclustParam}}.}

\item{leiden_params}{Leiden clustering parameters, passed to
\code{\link[igraph]{cluster_leiden}}.}
}
\value{
A data frame.
}
\description{
Moran's I has been computed for the same genes across different spatial
scales. This function gets the Moran's I values for all genes and resolutions
and clusters the patterns with which Moran's I's vary through scales, as a
way to cluster genes. Leiden and hierarchical clustering are performed on the
Moran's I values themselves and on the diff between adjacent scales. The
\code{\link[bluster]{approxSilhouette}} function can be used to assess
cluster quality.
}
