% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLeeSelect}
\alias{plotLeeSelect}
\title{Plot select Lee's L curves}
\usage{
plotLeeSelect(df_lees, lmm_res, pairs_use, title = NULL)
}
\arguments{
\item{df_lees}{Data frame with Lee's L results, same column names as in \code{df_res} in
\code{\link{runBinLMM}}.}

\item{lmm_res}{Linear mixed model results data frame}

\item{pairs_use}{Which pairs to plot}

\item{title}{Title of the plot}
}
\value{
A \code{ggplot2} oboject
}
\description{
Also annotate with relevant linear mixed model info
}
