% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMoranCurves}
\alias{plotMoranCurves}
\title{Plot Moran's I curves}
\usage{
plotMoranCurves(
  df,
  color_by = NULL,
  facet_by = NULL,
  show_null = FALSE,
  sfes = NULL,
  color_name = NULL
)
}
\arguments{
\item{df}{Data frame output from \code{\link{clusterMoranCurves}}, or any
data frame with a column named "gene" for genes, "side" for bin size as in
side length, and "moran" for Moran's I values, and optionally other
categorical columns indicating cluster membership of genes.}

\item{color_by}{A data frame with column "gene" for gene symbols or IDs and
one other column for color. Alternatively, name of a column in \code{df} to
use for coloring.}

\item{facet_by}{Name of a categorical or integer column in \code{df} to facet
the plot, not tidyeval.}

\item{show_null}{Logical, whether to show expected values of Moran's I under
null hypothesis (values are randomly permuted in space) and the interval
within which the null is not to be rejected (p >= 0.05 after Bonferroni
correction accounting for the number of genes and number of sides).}

\item{sfes}{The list of SFE objects from which \code{df} was made. Only
required when \code{show_null = TRUE}, to extract the number of bins in
order to compute the mean and variance of Moran's I for each bin size.}

\item{color_name}{Name to show for colors in the legend.}
}
\value{
A \code{ggplot} object.
}
\description{
The curves show how Moran's I changes over spatial scale, as indicated by bin
size used to aggregate the data.
}
