% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSFEs}
\alias{plotSFEs}
\title{Plot gene expression from multiple SFE objects}
\usage{
plotSFEs(
  sfes,
  feature,
  bbox = NULL,
  show_sizes = TRUE,
  ncol = NULL,
  nrow = NULL,
  widths = NULL,
  heights = NULL,
  design = NULL,
  ...
)
}
\arguments{
\item{sfes}{A list of SFE objects.}

\item{feature}{Gene to plot, only one gene}

\item{bbox}{Bounding box to plot a subarea}

\item{...}{Other arguments to pass to
\code{\link[Voyager]{plotSpatialFeature}}}
}
\value{
A \code{patchwork} object
}
\description{
This is useful to show what this gene looks like across different bin sizes.
When there are more than 200,000 cells or bins and no bbox is specified, then
scattermore is used to speed up plotting.
}
