% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSFEsBiscale}
\alias{plotSFEsBiscale}
\title{Plot pairs of genes in space from multiple SFE objects}
\usage{
plotSFEsBiscale(
  sfes,
  feature1,
  feature2,
  df_lee,
  colGeometryName = 1L,
  bbox = NULL,
  palette = "BlueGold",
  exprs_value = "logcounts",
  swap_rownames = NULL,
  title = NULL,
  crop = TRUE,
  show_sizes = TRUE,
  sep = "_",
  ncol = NULL,
  nrow = NULL,
  samples_use = NULL,
  side_use = NULL,
  widths = NULL,
  heights = NULL,
  design = NULL,
  legend_index = NULL
)
}
\arguments{
\item{sfes}{A list of SFE objects, whose names must be the bin sizes.}

\item{feature1}{First feature to plot}

\item{feature2}{Second feature to plot}

\item{df_lee}{Data frame from \code{\link{clusterLeeCurves}}, where Lee's L
values are extracted to annotate the plot.}

\item{colGeometryName}{Name of a \code{colGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute the metric. Use
\code{\link{colGeometryNames}} to look up names of the \code{sf} data
frames associated with cells/spots.}

\item{bbox}{Named numeric vector specifying a bounding box, either to zoom
into a smaller area or to show the box itself. The names should be "xmin",
"xmax", "ymin", and "ymax" in any order. The same bbox will be used for all
SFE objects in \code{sfes}.}

\item{palette}{Name of bivariate palette; see \code{\link[biscale]{bi_pal}}
for complete list of built-in palettes in the `biscale` package.}

\item{swap_rownames}{Column name of \code{rowData(object)} to be used to
identify features instead of \code{rownames(object)} when labeling plot
elements. If not found in \code{rowData}, then rownames of the gene count
matrix will be used.}

\item{crop}{Logical, whether to crop the samples by \code{bbox}. If
\code{FALSE}, then the \code{bbox} will be shown as a box on the plot to
indicate its location.}

\item{sep}{Separator between the two gene symbols in \code{df_lee}}

\item{ncol}{Number of columns if plotting multiple features. Defaults to
\code{NULL}, which means using the same logic as \code{facet_wrap}, which
is used by \code{patchwork}'s \code{\link{wrap_plots}} by default.}

\item{widths, heights}{The relative widths and heights of each column and row
in the grid. Will get repeated to match the dimensions of the grid. The
special value of \code{NA}/\verb{-1null} will behave as \verb{1null} unless a fixed aspect
plot is inserted in which case it will allow the dimension to expand or
contract to match the aspect ratio of the content}

\item{design}{Specification of the location of areas in the layout. Can either
be specified as a text string or by concatenating calls to \code{\link[patchwork:area]{area()}} together.
See the examples for further information on use.}

\item{legend_index}{Index of position to place the legend. If \code{NULL},
then the legend will not be shown. Unlike in univariate plots, the legend
itself is a \code{ggplot2} object}
}
\value{
A \code{patchwork} object.
}
\description{
This function is useful to visualize pairs of genes whose Lee's L is of
interest.
}
