% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{plotSFEsRedDim}
\alias{plotSFEsRedDim}
\title{Plot dimension reduction in space for multiple SFE objects}
\usage{
plotSFEsRedDim(
  sfes,
  pc = 1,
  reduction = "PCA",
  field = "rotation",
  divergent = TRUE,
  diverge_center = 0,
  bbox = NULL
)
}
\arguments{
\item{sfes}{A list of SFE objects, for which the dimension reduction of
interest has been computed.}

\item{pc}{Index of principal component or eigenvector or loading vector to
plot. Facetting is used when length of this argument is greater than 1.}

\item{reduction}{Name of dimension reduction to plot}

\item{field}{Field in the attribute of the dimension reduction where variance
explained or eigenvalues are stored. If this field is not found, then the
components will not be flipped and flipping is not necessary for dimension
reductions such as NMF which only produce non-negative results.}
}
\value{
A \code{patchwork} object
}
\description{
This function not only plots dimension reductions in space, but also aligns
the components for a consistent look because in PCA, the PC's can be flipped
in sign.
}
