% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{runBinAnalyses}
\alias{runBinAnalyses}
\title{Run basic analyses on aggregated SFE objects of all bin sizes}
\usage{
runBinAnalyses(
  dir,
  out_dir,
  tissue_geometry,
  min_props = 0.9,
  quantiles = NULL,
  ncomponents = 30,
  queen = FALSE,
  zero.policy = TRUE,
  p.adjust.method = "BH",
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{dir}{Directory where the binned outputs are located. Output of each bin
size must be in a directory named "binx", such as "bin12" for 12 micron
bins.}

\item{tissue_geometry}{Either tissue boundary (with holes if present in
tissue) or cell segmentation polygons. Area of overlap of each bin with
this geometry will be computed.}

\item{min_props}{Minimum proportion of each bin overlapping tissue; bins that
don't overlap enough will be removed to mitigate edge effect. This should
be a numeric vector of the same length as the number of bin sizes and can
differ for different bin sizes. If length 1, then the same value will be
applied to all bin sizes. Otherwise, it must have names corresponding to
bin sizes.}

\item{quantiles}{Numeric vector of quantiles of area of bin overlapping
tissue, can differ for different bin sizes. If not NULL, this will
supercede \code{min_props}.}

\item{ncomponents}{Number of components to compute for PCA.}

\item{queen}{if TRUE, a single shared boundary point meets the contiguity condition, if FALSE, more than one shared point is required; note that more than one shared boundary point does not necessarily mean a shared boundary line}

\item{zero.policy}{default \code{attr(listw, "zero.policy")} as set when \code{listw} was created, if attribute not set, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA}

\item{p.adjust.method}{Method to correct for multiple testing, passed to
\code{\link[spdep]{p.adjustSP}}. Methods allowed are in
\code{\link{p.adjust.methods}}.}

\item{BPPARAM}{A \code{\link{BiocParallelParam}} object specifying whether
and how computing the metric for numerous genes shall be parallelized.}

\item{...}{Other arguments passed to S4 method (for convenience wrappers like
\code{calculateMoransI}) or method used to compute metrics as specified by
the argument \code{type} (as in more general functions like
\code{calculateUnivariate}). See documentation of functions with the same
name as specified in \code{type} in the \code{spdep} package for the method
specific arguments. For variograms, see \code{\link[Voyager]{.variogram}}.}
}
\value{
Invisibly \code{out_dir}; the log normalization, QC, PCA, adjacency
  graph, and Moran's I will be stored in the SFE object; the SFE object with
  the results will be written to \code{out_dir} with directory names
  "binx_esda". A data frame for Moran's I and  Lee's L across bin sizes will
  be written to \code{out_dir} as CSV files.
}
\description{
This function runs log normalization, bin-level QC, PCA, adjacency graph,
Moran's I, and Lee's L for all binned SFE objects in a directory, given that
bins overlapping too little with tissue have been removed.
}
